GRAL: "Fixed detailfragment video and enable navigation detail to itself"

+++ app/src/main/java/com/cbs/demospike/MainFragment.kt
@@ -18,7 +18,6 @@ import androidx.leanback.app.BrowseFragment
 import androidx.leanback.app.BrowseSupportFragment
 import androidx.leanback.widget.*
 import androidx.navigation.findNavController
-import androidx.navigation.fragment.findNavController
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.resource.drawable.GlideDrawable
 import com.bumptech.glide.request.animation.GlideAnimation
@@ -123,7 +122,7 @@ class MainFragment : BrowseSupportFragment() {
             row: Row
         ) {
             if (item is Movie) {
-                val action = MainFragmentDirections.actionMainToDetail(item.id.toInt())
+                val action = MainFragmentDirections.actionMainToDetail(item.id)
                 view?.findNavController()?.navigate(action)
             } else if (item is String) {
                 if (item.contains(getString(R.string.error_fragment))) {
